requirements.txt
flask
pdfminer.six
chromadb
celery
SQLAlchemy
chatwrap @ git+https://github.com/stefanbertici/chatwrap.git
